serde: move async functions to async.h
